migrate filters to subcomponents
