+++ src/components/FilterPanel/ZipCodeFilter.js
@@ -1,0 +1,44 @@
+import FormControl from "@material-ui/core/FormControl";
+import FormGroup from "@material-ui/core/FormGroup";
+import Input from "@material-ui/core/Input";
+import FormControlLabel from "@material-ui/core/FormControlLabel";
+import Cookies from "universal-cookie";
+
+const cookies = new Cookies();
+export function getZipCodeCookie() {
+    const savedZip = cookies.get("ZIPCode");
+    return savedZip ? savedZip : "";
+}
+export default function ZipCodeFilter(props) {
+    const handleChange = (e) => {
+        const targetZip = e.target.value;
+        const zipValid = targetZip === "" || targetZip.match(/\d{5}/);
+        /* TODO: we have a mix of cookie stuff in and out of this component.
+         * It should all be in one place. */
+        if (zipValid) {
+            cookies.set("ZIPCode", targetZip, { path: "/" });
+        }
+        props.onChange(targetZip);
+    };
+
+    return (
+        <FormControl component="fieldset" className={props.className}>
+            {/*<FormLabel component="legend"></FormLabel>*/}
+            <FormGroup>
+                <FormControlLabel
+                    control={
+                        <Input
+                            value={props.zipCode}
+                            onChange={handleChange}
+                            name="zipCode"
+                            inputProps={{ "data-testid": "zip-input" }}
+                        />
+                    }
+                    label="ZIP Code"
+                    labelPlacement="top"
+                    aria-label="ZIP Code"
+                />
+            </FormGroup>
+        </FormControl>
+    );
+}
